Move classes in common into packages
