--- common/src/main/java/nl/tudelft/wdm/group1/common/OrderItemId.java
@@ -1,41 +1,0 @@
-package nl.tudelft.wdm.group1.common;
-
-import javax.persistence.Entity;
-import javax.persistence.Id;
-import java.util.Objects;
-import java.util.UUID;
-
-@Entity
-public class OrderItemId {
-    @Id
-    private UUID id;
-
-    public OrderItemId() {
-
-    }
-
-    public OrderItemId(UUID id) {
-        this.id = id;
-    }
-
-    public UUID getId() {
-        return id;
-    }
-
-    public void setId(UUID id) {
-        this.id = id;
-    }
-
-    @Override
-    public boolean equals(Object o) {
-        if (this == o) return true;
-        if (o == null || getClass() != o.getClass()) return false;
-        OrderItemId that = (OrderItemId) o;
-        return Objects.equals(id, that.id);
-    }
-
-    @Override
-    public int hashCode() {
-        return Objects.hash(id);
-    }
-}
