+++ common/src/main/java/nl/tudelft/wdm/group1/common/exception/InsufficientStockException.java
@@ -1,11 +1,11 @@
-package nl.tudelft.wdm.group1.common;
+package nl.tudelft.wdm.group1.common.exception;
 
 import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.ResponseStatus;
 
 @ResponseStatus(value = HttpStatus.NOT_FOUND)
-public class ResourceNotFoundException extends Exception {
-    public ResourceNotFoundException(String message) {
+public class InsufficientStockException extends Exception {
+    public InsufficientStockException(String message) {
         super(message);
     }
 }
