+++ common/src/main/java/nl/tudelft/wdm/group1/common/model/Payment.java
@@ -1,4 +1,4 @@
-package nl.tudelft.wdm.group1.common;
+package nl.tudelft.wdm.group1.common.model;
 
 import javax.persistence.Entity;
 import javax.persistence.Id;
