+++ service/payments/src/main/java/nl/tudelft/wdm/group1/payments/events/OrderConsumer.java
@@ -1,8 +1,8 @@
 package nl.tudelft.wdm.group1.payments.events;
 
-import nl.tudelft.wdm.group1.common.OrdersTopics;
-import nl.tudelft.wdm.group1.common.Payment;
-import nl.tudelft.wdm.group1.common.Order;
+import nl.tudelft.wdm.group1.common.model.Order;
+import nl.tudelft.wdm.group1.common.model.Payment;
+import nl.tudelft.wdm.group1.common.topics.OrdersTopics;
 import nl.tudelft.wdm.group1.payments.PaymentRepository;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
