+++ service/payments/src/test/java/nl/tudelft/wdm/group1/payments/PaymentsApplicationTest.java
@@ -1,11 +1,11 @@
 package nl.tudelft.wdm.group1.payments;
 
 import nl.tudelft.wdm.group1.common.KafkaResponse;
-import nl.tudelft.wdm.group1.common.Payment;
-import nl.tudelft.wdm.group1.common.RestTopics;
+import nl.tudelft.wdm.group1.common.model.Payment;
 import nl.tudelft.wdm.group1.common.payload.PaymentAddPayload;
 import nl.tudelft.wdm.group1.common.payload.PaymentDeletePayload;
 import nl.tudelft.wdm.group1.common.payload.PaymentGetPayload;
+import nl.tudelft.wdm.group1.common.topics.RestTopics;
 import org.apache.kafka.clients.consumer.Consumer;
 import org.apache.kafka.clients.consumer.ConsumerConfig;
 import org.apache.kafka.clients.producer.Producer;
