+++ service/payments/src/test/java/nl/tudelft/wdm/group1/payments/events/ConsumerTest.java
@@ -1,13 +1,13 @@
 package nl.tudelft.wdm.group1.payments.events;
 
-import nl.tudelft.wdm.group1.common.*;
+import nl.tudelft.wdm.group1.common.model.Order;
+import nl.tudelft.wdm.group1.common.model.Payment;
 import nl.tudelft.wdm.group1.payments.PaymentRepository;
 import org.junit.Before;
 import org.junit.Test;
 
 import java.util.UUID;
 
-import static org.assertj.core.api.Assertions.assertThat;
 import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.verify;
