+++ service/rest/src/main/java/nl/tudelft/wdm/group1/rest/events/KafkaInteraction.java
@@ -2,8 +2,8 @@ package nl.tudelft.wdm.group1.rest.events;
 
 import nl.tudelft.wdm.group1.common.KafkaErrorResponse;
 import nl.tudelft.wdm.group1.common.KafkaResponse;
-import nl.tudelft.wdm.group1.common.RestTopics;
 import nl.tudelft.wdm.group1.common.payload.RestPayload;
+import nl.tudelft.wdm.group1.common.topics.RestTopics;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
