+++ service/stock/src/main/java/nl/tudelft/wdm/group1/stock/events/Producer.java
@@ -1,9 +1,9 @@
 package nl.tudelft.wdm.group1.stock.events;
 
-import nl.tudelft.wdm.group1.common.Order;
-import nl.tudelft.wdm.group1.common.OrdersTopics;
-import nl.tudelft.wdm.group1.common.StockItem;
-import nl.tudelft.wdm.group1.common.StockTopics;
+import nl.tudelft.wdm.group1.common.model.Order;
+import nl.tudelft.wdm.group1.common.model.StockItem;
+import nl.tudelft.wdm.group1.common.topics.OrdersTopics;
+import nl.tudelft.wdm.group1.common.topics.StockTopics;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
