+++ service/users/src/main/java/nl/tudelft/wdm/group1/users/UserRepository.java
@@ -1,7 +1,7 @@
 package nl.tudelft.wdm.group1.users;
 
-import nl.tudelft.wdm.group1.common.ResourceNotFoundException;
-import nl.tudelft.wdm.group1.common.User;
+import nl.tudelft.wdm.group1.common.exception.ResourceNotFoundException;
+import nl.tudelft.wdm.group1.common.model.User;
 import org.springframework.data.repository.CrudRepository;
 
 import java.util.UUID;
